requirements.txt
# CyberSec Dashboard — Python Dependencies

# Web / WebSocket
aiohttp==3.9.5
websockets==12.0

# Database
aiomysql==0.2.0
PyMySQL==1.1.1

# AI / ML
scikit-learn==1.4.2
numpy==1.26.4

# Auth
PyJWT==2.8.0
bcrypt==4.1.3

# HTTP client (GeoIP)
aiohttp==3.9.5

# Config
python-dotenv==1.0.1
